Doing some speed testing.
The time measurements in the scripts just tells me in-game time,
which doesn't advance at all when the scripts don't yield, so
I'm just using it to figure out if I messed up and allowed the scripts
to yield.

When the scripts are autoyielding it messes up timing because the
lua updates dominate the timing.

I'm actually measuring this with a stopwatch on my phone for convenience, lol,

but I'm only after rough estimates and using large enough examples to
make it easily measurable by a human pressing a stopwatch button.

+++ Kerbalua/Kerbalua/Scripting/KerbaluaScript.cs
@@ -23,7 +23,7 @@ namespace Kerbalua.Scripting
 		private int execlimit = defaultExecLimit;
 		private const int defaultExecLimit=500;
 		private const int execLimitMin=100;
-		private const int execLimitMax=1000;
+		private const int execLimitMax=1000000000;
 		static KerbaluaScript _instance;
 		public static KerbaluaScript Instance
 		{
@@ -149,9 +149,13 @@ namespace Kerbalua.Scripting
 		public void Terminate()
 		{
 			coroutine = null;
-			process.shutdown -= Terminate;
-			process.terminate();
-			process = null;
+			if (process!=null)
+			{
+				process.shutdown -= Terminate;
+				process.terminate();
+				process = null;
+			}
+
 			sleepwatch.Reset();
 			sleeptimeMillis=0;
 		}
